FASTAPI/requirements.txt
#inorder to download everything at once in the terminal just do 
# pip install -r requirements.txt

fastapi
uvicorn
sqlalchemy
PyJWT
passlib
bcrypt
python-jose
python-multipart
